Coding - Scope remaining ~35 Visualization enums as enum class
Converts unscoped enums across Visualization subdirs to enum class:
MeshVS (1 of 3: MeshSelectionMethod), Font (1 of 4: StrictLevel),
Image (none — all reverted), WNT, V3d (TypeOfView/Visualization/
StereoDumpOptions), StdSelect (all 3), SelectMgr (6 of 7),
Prs3d (7 of 10), DsgPrs, StdPrs, PrsMgr (both), PrsDim (7 of 8),
OpenGl (2 of 4: MaterialFlag, LayerFilter).

Kept as plain enum because of pervasive int interoperation, bitflag
semantics, or array-index usage:
- Image_Format, Image_CompressedFormat (array indices, GL enums)
- Font_FontAspect, Font_Hinting (bitflag), Font_UnicodeSubset
  (array index)
- V3d_TypeOfOrientation, V3d_TypeOfAxe (array indices in Trihedron
  and view-manipulation helpers)
- OpenGl_FeatureFlag (converted to bool via !val pattern),
  OpenGl_ShaderProgramDumpLevel (converted to bool)
- SelectMgr_SelectionType, Select3D_TypeOfSensitivity (passed as
  int through selection pipeline)
- Prs3d_DatumAxes (bitflag), Prs3d_DatumAttribute (array index),
  Prs3d_TypeOfHighlight (array size/index in AIS_InteractiveContext)
- PrsDim_DimensionSelectionMode (passed as int through
  SelectionManager)
- MeshVS_DrawerAttribute (passed as int to drawer GetInteger/
  SetInteger), MeshVS_EntityType (bitflag), also
  MeshVS_SelectionModeFlags (bitflag, excluded up-front)

+++ src/Visualization/TKV3d/Select3D/Select3D_InteriorSensitivePointSet.cxx
@@ -256,7 +256,9 @@ bool Select3D_InteriorSensitivePointSet::overlapsElement(
   int                                            aPolygIdx = myPolygonsIdxs->Value(theElemIdx);
   const occ::handle<Select3D_SensitivePoly>&     aPolygon  = myPlanarPolygons.Value(aPolygIdx);
   const occ::handle<NCollection_HArray1<gp_Pnt>> aPoints   = aPolygon->Points3D();
-  return theMgr.OverlapsPolygon(aPoints->Array1(), Select3D_TOS_INTERIOR, thePickResult);
+  return theMgr.OverlapsPolygon(aPoints->Array1(),
+                                static_cast<int>(Select3D_TOS_INTERIOR),
+                                thePickResult);
 }
 
 //=================================================================================================
